+++ nj-derive/src/ast/types.rs
@@ -7,6 +7,9 @@ use syn::Result;
 use syn::spanned::Spanned;
 use syn::DeriveInput;
 use syn::Data;
+use syn::DataEnum;
+use syn::DataStruct;
+use syn::Variant;
 use syn::Fields;
 use syn::GenericParam;
 use syn::WhereClause;
@@ -108,22 +111,68 @@ impl MyTupleType<'_> {
 }
 
 #[derive(Debug)]
-pub enum MyStruct<'a> {
-    Named {
-        name: &'a Ident,
-        fields: Vec<MyField<'a>>,
-        generics: MyGenerics<'a>,
-    },
-    Unnamed {
-        name: &'a Ident,
-        fields: Vec<MyFieldType<'a>>,
-        generics: MyGenerics<'a>,
-    },
+pub struct MyDeriveInput<'a> {
+    pub name: &'a Ident,
+    pub generics: MyGenerics<'a>,
+    pub payload: MyDerivePayload<'a>,
 }
 
 #[derive(Debug)]
-pub struct MyField<'a> {
+pub enum MyDerivePayload<'a> {
+    Struct(MyStruct<'a>),
+    Enum(MyEnum<'a>),
+}
+
+impl<'a> MyDeriveInput<'a> {
+    pub fn from_ast(input: &'a DeriveInput) -> Result<MyDeriveInput> {
+        let name = &input.ident;
+        let generic_params = input.generics.params.clone().into_iter().collect();
+        let generics = MyGenerics {
+            params: generic_params,
+            where_clause: &input.generics.where_clause,
+        };
+
+        match &input.data {
+            Data::Struct(inner_struct) => {
+                let parsed_struct = MyStruct::from_ast(&inner_struct)?;
+                Ok(MyDeriveInput {
+                    name,
+                    generics,
+                    payload: MyDerivePayload::Struct(parsed_struct),
+                })
+            }
+            Data::Enum(inner_enum) => {
+                let parsed_enum = MyEnum::from_ast(&inner_enum)?;
+                Ok(MyDeriveInput {
+                    name,
+                    generics,
+                    payload: MyDerivePayload::Enum(parsed_enum),
+                })
+            }
+            Data::Union(_) => Err(Error::new(
+                input.span(),
+                "Unions are not supported \
+                for automatic conversion to JavaScript representation",
+            )),
+        }
+    }
+}
+
+#[derive(Debug)]
+pub enum MyFields<'a> {
+    Named(Vec<MyNamedField<'a>>),
+    Unnamed(Vec<MyUnnamedField<'a>>),
+    Unit,
+}
+
+#[derive(Debug)]
+pub struct MyNamedField<'a> {
     pub name: &'a Ident,
+    pub ty: MyFieldType<'a>,
+}
+
+#[derive(Debug)]
+pub struct MyUnnamedField<'a> {
     pub ty: MyFieldType<'a>,
 }
 
@@ -133,14 +182,40 @@ pub enum MyFieldType<'a> {
     Ref(MyReferenceType<'a>),
 }
 
-#[derive(Debug)]
-pub struct MyGenerics<'a> {
-    pub params: Vec<GenericParam>,
-    pub where_clause: &'a Option<WhereClause>,
+impl<'a> MyFields<'a> {
+    pub fn from_ast(input: &'a Fields) -> Result<MyFields> {
+        match &input {
+            Fields::Named(named_fields) => {
+                let fields = named_fields
+                    .named
+                    .iter()
+                    .filter_map(|field| field.ident.as_ref().map(|ident| (ident, &field.ty)))
+                    .map(|(ident, ty)| {
+                        MyFieldType::from(&ty).map(|ty| MyNamedField { name: &ident, ty })
+                    })
+                    .collect::<Result<Vec<MyNamedField<'a>>>>()?;
+
+                Ok(MyFields::Named(fields))
+            }
+            Fields::Unnamed(unnamed_fields) => {
+                let fields = unnamed_fields
+                    .unnamed
+                    .iter()
+                    .map(|field| MyFieldType::from(&field.ty))
+                    .collect::<Result<Vec<MyFieldType<'a>>>>()?
+                    .into_iter()
+                    .map(|ty| MyUnnamedField { ty })
+                    .collect::<Vec<MyUnnamedField<'a>>>();
+
+                Ok(MyFields::Unnamed(fields))
+            }
+            Fields::Unit => Ok(MyFields::Unit),
+        }
+    }
 }
 
 impl<'a> MyFieldType<'a> {
-    pub fn from(ty: &'a Type) -> Result<Self> {
+    pub fn from(ty: &'a Type) -> Result<MyFieldType> {
         match ty {
             Type::Path(type_path) => Ok(MyFieldType::Path(MyTypePath::from(type_path)?)),
             Type::Reference(reference) => Ok(MyFieldType::Ref(MyReferenceType::from(reference)?)),
@@ -153,63 +228,55 @@ impl<'a> MyFieldType<'a> {
     }
 }
 
-impl<'a> MyStruct<'a> {
-    pub fn from_ast(input: &'a DeriveInput) -> Result<MyStruct> {
-        let struct_data = match &input.data {
-            Data::Struct(inner_struct) => Ok(inner_struct),
-            Data::Enum(_) => Err(Error::new(
-                input.span(),
-                "Enums are not supported \
-                for automatic conversion to JavaScript representation",
-            )),
-            Data::Union(_) => Err(Error::new(
-                input.span(),
-                "Unions are not supported \
-                for automatic conversion to JavaScript representation",
-            )),
-        }?;
+#[derive(Debug)]
+pub struct MyEnum<'a> {
+    pub variants: Vec<MyVariant<'a>>,
+}
 
-        let generic_params = input.generics.params.clone().into_iter().collect();
-        let generics = MyGenerics {
-            params: generic_params,
-            where_clause: &input.generics.where_clause,
-        };
+impl<'a> MyEnum<'a> {
+    pub fn from_ast(enum_data: &'a DataEnum) -> Result<MyEnum> {
+        let variants = enum_data
+            .variants
+            .iter()
+            .map(|v| MyVariant::from_ast(v))
+            .collect::<Result<Vec<MyVariant>>>()?;
 
-        match &struct_data.fields {
-            Fields::Named(named_fields) => {
-                let fields = named_fields
-                    .named
-                    .iter()
-                    .filter_map(|field| field.ident.as_ref().map(|ident| (ident, &field.ty)))
-                    .map(|(ident, ty)| {
-                        MyFieldType::from(&ty).map(|ty| MyField { name: &ident, ty })
-                    })
-                    .collect::<Result<Vec<MyField<'a>>>>()?;
-
-                Ok(MyStruct::Named {
-                    name: &input.ident,
-                    fields,
-                    generics,
-                })
-            }
-            Fields::Unnamed(unnamed_fields) => {
-                let fields = unnamed_fields
-                    .unnamed
-                    .iter()
-                    .map(|field| MyFieldType::from(&field.ty))
-                    .collect::<Result<Vec<MyFieldType<'a>>>>()?;
-
-                Ok(MyStruct::Unnamed {
-                    name: &input.ident,
-                    fields,
-                    generics,
-                })
-            }
-            Fields::Unit => Err(Error::new(
-                input.span(),
-                "Unit structs are not supported for automatic conversion yet. \
-                If you would like to see them supported, please reach out.",
-            )),
-        }
+        Ok(MyEnum { variants })
     }
+}
+
+#[derive(Debug)]
+pub struct MyVariant<'a> {
+    pub name: &'a Ident,
+    pub fields: MyFields<'a>,
+}
+
+impl<'a> MyVariant<'a> {
+    pub fn from_ast(variant_data: &'a Variant) -> Result<MyVariant> {
+        let fields = MyFields::from_ast(&variant_data.fields)?;
+
+        Ok(MyVariant {
+            name: &variant_data.ident,
+            fields,
+        })
+    }
+}
+
+#[derive(Debug)]
+pub struct MyStruct<'a> {
+    pub fields: MyFields<'a>,
+}
+
+impl<'a> MyStruct<'a> {
+    pub fn from_ast(struct_data: &'a DataStruct) -> Result<MyStruct> {
+        let fields = MyFields::from_ast(&struct_data.fields)?;
+
+        Ok(MyStruct { fields })
+    }
+}
+
+#[derive(Debug)]
+pub struct MyGenerics<'a> {
+    pub params: Vec<GenericParam>,
+    pub where_clause: &'a Option<WhereClause>,
 }