+++ nj-derive/ui-tests/fail_enum.rs
@@ -1,7 +1,0 @@
-use node_bindgen::derive::node_bindgen;
-
-#[node_bindgen]
-enum TestEnum {
-    Something,
-    Else
-}